Make Error an interface rather than a concrete type
In Go, `error` is an interface, so making `errors.Error` a concrete type
broke code such as the following:

	func goError() error {
	    return nil
	}

	func myError() *errors.Error {
	    return nil
	}

	func main() {

	    err := goError()
	    if err != nil {
	        fmt.Println("goError not nil")
	    }

	    err = myError()
	    if err != nil {
	        fmt.Println("myError not nil")
	    }

	}

The code above prints "myError not nil", which can be troubling.
This is due to the way interfaces work in go
(see https://news.ycombinator.com/item?id=5196241 for reference).

Introducing an `Error` interface is a breaking change, but resolves
this kind of mistake by unexperienced go developpers.

In this commit, I have:
- renamed `Error` to `wrappedError` (not exported)
- added a new `Error` interface with the existing methods
- Added an `Underlying()` method to replace the `Err` property that
  cannot be accessed on the interface